Implementado Comando de Update de Venda

+++ PaymentAPI.Domain/Features/Item.cs
@@ -5,7 +5,7 @@ namespace PaymentAPI.Domain.Features
     public class Item
     {
         public int ItemId { get; set; }
-        public int SaleId { get; set; }
+        public long SaleId { get; set; }
         public string Name { get; set; }
         public decimal Price { get; set; }
         public int Quantity { get; set; }
